requirements.txt
accelerate==0.33.0
altair==5.3.0
attrs==23.2.0
bitsandbytes==0.43.2
blinker==1.8.2 
cachetools==5.4.0 
certifi==2024.7.4
charset-normalizer==3.3.2 
click==8.1.7 
easyocr==1.7.1 
einops==0.8.0 
filelock==3.15.4
flash-attn==2.6.2 
fsspec==2024.6.1
gitdb==4.0.11
gitpython==3.1.43
huggingface-hub==0.24.2
idna==3.7   
imageio==2.34.2
jinja2==3.1.4 
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lazy-loader==0.4   
markdown-it-py==3.0.0 
markupsafe==2.1.5 
mdurl==0.1.2 
mpmath==1.3.0 
networkx==3.3   
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82 
nvidia-nvtx-cu12==12.1.105
opencv-python-headless==4.10.0.84
packaging==24.1  
pandas==2.2.2 
pillow==10.4.0
pip==24.0
protobuf==5.27.2
psutil==6.0.0 
pyarrow==17.0.0
pyclipper==1.3.0.post5
pydeck==0.9.1 
pygments==2.18.0
python-bidi==0.5.1 
python-dateutil==2.9.0.post0
pytz==2024.1
pyyaml==6.0.1 
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
rpds-py==0.19.1
safetensors==0.4.3 
scikit-image==0.24.0
scipy==1.14.0
setuptools==71.0.4
shapely==2.0.5 
six==1.16.0
smmap==5.0.1 
streamlit==1.37.0
sympy==1.13.1
tenacity==8.5.0 
tifffile==2024.7.24
timm==1.0.7 
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.4.0 
torchvision==0.19.0
tornado==6.4.1 
tqdm==4.66.4
transformers==4.43.3
triton==3.0.0 
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2 
watchdog==4.0.1 
wheel==0.43.0